requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.9.4
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
aspose-pdf==24.4.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
awscrt==0.19.19
Babel==2.14.0
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
boto3==1.34.84
botocore==1.34.84
cachetools==5.3.3
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
clr-loader==0.2.6
comm==0.2.2
configobj==5.0.8
configparser==7.0.0
contourpy==1.2.1
cryptography==40.0.2
cycler==0.12.1
databutton==0.38.12
dataclasses-json==0.6.6
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
easyocr==1.7.1
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.0
executing==2.0.1
faiss-gpu==1.7.2
fastapi==0.110.1
fastjsonschema==2.19.1
ffmpy==0.3.2
filelock==3.13.4
fire==0.6.0
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.1.3
glob2==0.7
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gradio==4.31.5
gradio_client==0.16.4
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
imageio==2.34.1
import-ipynb==0.1.4
importlib_resources==6.4.0
inflection==0.5.1
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
IronPdf==2024.4.1.2
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.0.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.26.0
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.3
langchain-openai==0.1.8
langchain-text-splitters==0.2.0
langsmith==0.1.67
lazy_loader==0.4
lightning==2.2.2
lightning-cloud==0.5.64
lightning-utilities==0.10.1
lightning_sdk==0.1.4
litdata==0.2.2
llamaapi==0.1.36
looseversion==1.3.0
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
ninja==1.11.1.1
nipype==1.8.6
notebook_shim==0.2.4
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
objprint==0.2.3
openai==1.31.1
openapi==1.1.0
opencv-python-headless==4.9.0.80
openpyxl==3.1.3
orjson==3.10.3
overrides==7.7.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.23.4
prov==2.0.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pydot==2.0.0
pydotenv==0.0.7
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.8.0
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
pypdfium2==4.29.0
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pythonnet==3.0.3
pytorch-lightning==2.2.2
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
pyzmq==25.1.2
rdflib==7.0.0
referencing==0.34.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
ruff==0.4.5
s3transfer==0.10.1
scikit-image==0.23.2
scikit-learn==1.3.2
scipy==1.11.4
semantic-version==2.10.0
Send2Trash==1.8.3
shapely==2.0.4
shellingham==1.5.4
simple-term-menu==1.6.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tenacity==8.3.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.4.0
tifffile==2024.5.3
tiktoken==0.7.0
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.1+cu121
torchmetrics==1.3.1
torchvision==0.17.1+cu121
tornado==6.4
tqdm==4.66.1
traitlets==5.14.2
traits==6.3.2
triton==2.2.0
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.29.0
viztracer==0.16.2
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.2
widgetsnbextension==4.0.10
xgboost==2.0.2
yarl==1.9.4
